Use bulk percolation as default API int YPercolateService
The default API (ether both default and batch percolators) uses the
single request percolation API. The default percolator creates the entire
environment per request from the batch. The batch percolator just embed
the documents into the single requests.

YPercolator makes the bulk percolation method as default: requests bulk
in and responses bulk out.

That method causes some unexpected failures when unhandled exceptions are
thrown and responses bulk is not completed. This causes ES to stuck on
returning the response data to the caller and stops in 60 seconds timeout.